ao_coreaudio: skip channel labels marked with and unknown tag
I don't think this is really a very good idea because I think it is
conceptually incorrect but other multimedia software in uses this approach
(VLC and xbmc), and it seems to make the conversion more robust in certain
cases.

For example it has been reported, that configuring a receiver that can output
7.1 to output 5.1, will make CoreAudio report 8 channel descriptions, and the
last 2 descriptions withh be tagged kAudioChannelLabel_Unknown.